UDP: - Actualización para la captura de la tolerance (en formato double).

+++ include/DataLoader.h
@@ -7,7 +7,9 @@
 #include "ABCMethod.h" // Para la definición de SKUData
 
 SKUData loadSKUData(const std::string& filename);
+
 std::map<std::string, double> loadNormalizedFeatures(const std::string& filename);
-void loadSimulationConfig(const std::string& filename, int& numberOfIterations, int& tolerance, int& daysToSimulate);
+
+void loadSimulationConfig(const std::string& filename, int& numberOfIterations, double& tolerance, int& daysToSimulate);
 
 #endif // DATALOADER_H
